Repackaged the project, moved resourceconfiguration package to project root
