+++ src/main/java/gr/aegean/icsd/icarus/resourceconfiguration/ResourceConfiguration.java
@@ -1,4 +1,4 @@
-package gr.aegean.icsd.icarus.test.resourceconfiguration;
+package gr.aegean.icsd.icarus.resourceconfiguration;
 
 import gr.aegean.icsd.icarus.test.Test;
 import gr.aegean.icsd.icarus.util.aws.AwsRegion;
